set atlas datatype to int16 if it is not already

+++ ndmg/graph/gen_graph.py
@@ -196,7 +196,7 @@ class graph_tools(object):
         mx = len(np.unique(self.rois.astype(np.int64))) - 1
         self.g = nx.Graph(ecount=0, vcount=mx)
         edge_dict = defaultdict(int)
-        node_dict = dict(zip(np.unique(self.rois) + 1, np.arange(mx) + 1))
+        node_dict = dict(zip(np.unique(self.rois).astype('int16') + 1, np.arange(mx) + 1))
 
         # Add empty vertices
         for node in range(1, mx + 1):
@@ -221,7 +221,7 @@ class graph_tools(object):
 
             lab_arr = self.rois[i, j, k]
             endlabels = []
-            for lab in np.unique(lab_arr):
+            for lab in np.unique(lab_arr).astype('int16'):
                 if (lab > 0) and (np.sum(lab_arr == lab) >= overlap_thr):
                     endlabels.append(node_dict[lab])
